Allow items on the "me" side of a trade offer to be price checked
Items in a steam trade are added with the same HTML ID attribute. Having multiple elements with the same id is problematic for jQuery, when selecting them by using the "#id" selector. To ensure that the checked price span is added to the items in the trade offer window rather than only the inventory, I have added the classes of the parent element to each element.